# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Liketexts API
  version: 1.0.0
  description: 'Substack-style platform: readers subscribe to writers via allocation
    slots; posts are paywalled (preview + subscriber content).'
servers:
- url: http://localhost:5000
paths:
...
  /api/v1/writers:
    get:
      summary: Searchable list of public writers
      tags:
      - Writers
      parameters:
      - name: q
        in: query
        required: false
        description: Case-insensitive display-name filter.
        schema:
          type: string
      - name: limit
        in: query
        required: false
        description: Max writers to return.
        schema:
          type: integer
          default: 20
      - name: offset
        in: query
        required: false
        description: Writers to skip.
        schema:
          type: integer
          default: 0
      responses:
        '200':
          description: Writer catalog page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WritersList'
  /api/v1/writers/{writer_id}:
    get:
      summary: Writer profile and published newsletters
      tags:
      - Writers
      description: Posts are paywall-masked for readers without an allocation.
      parameters:
      - name: writer_id
        in: path
        required: true
        description: Writer id.
        schema:
          type: string
          format: uuid
      - &id001
        name: Authorization
        in: header
        required: false
        description: Optional `Bearer <access_token>`. Full content is returned when
          the reader has an active allocation to the writer.
        schema:
          type: string
      responses:
        '200':
          description: Writer with masked posts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WriterDetail'
        '400':
          description: Invalid writer_id format.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Writer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/writers/{writer_id}/follow:
    post:
      summary: Follow a writer (preview emails, no allocation).
      tags:
      - Writers
      parameters:
      - name: writer_id
        in: path
        required: true
        description: Writer id.
        schema:
          type: string
          format: uuid
      - *id001
      responses:
        '201':
          description: Now following.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowState'
        '400':
          description: Invalid writer_id or self-follow.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Writer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
    delete:
      summary: Stop following a writer.
      tags:
      - Writers
      parameters:
      - name: writer_id
        in: path
        required: true
        description: Writer id.
        schema:
          type: string
          format: uuid
      - *id001
      responses:
        '200':
          description: No longer following.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowState'
        '401':
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Writer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
...
  /api/v1/posts/writer:
    get:
      summary: Current writer's posts
      tags:
      - Publishing
      parameters:
      - name: status
        in: query
        required: false
        description: Filter by status.
        schema:
          type: string
          enum:
          - draft
          - scheduled
          - published
          - cancelled
      responses:
        '200':
          description: Own posts, unmasked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WriterPosts'
        '401':
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
...
components:
  schemas:
    Error:
      type: object
      required:
      - title
      - status
      - detail
      - message
      - code
      properties:
        title:
          type: string
          example: Unauthorized
        status:
          type: integer
          example: 401
        detail:
          type: string
        message:
          type: string
          description: Alias of detail.
        code:
          type: string
          example: UNAUTHORIZED
...
    PostView:
      type: object
      description: Paywall-applied post. `subscriber_content` is null without an active
        allocation to the writer (or authorship).
      required:
      - id
      - writer_id
      - title
      - preview_content
      - status
      - published_at
      - created_at
      - subscriber_content
      - has_full_access
      - writer_name
      - writer_avatar_url
      properties:
        id:
          type: string
          format: uuid
        writer_id:
          type: string
          format: uuid
        title:
          type: string
        preview_content:
          type: string
        status:
          type: string
          example: published
        published_at: &id003
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        subscriber_content: *id002
        has_full_access:
          type: boolean
        writer_name: *id002
        writer_avatar_url: *id002
...
    WriterPost:
      type: object
      description: Full post for the owning writer (no paywall masking).
      required:
      - id
      - title
      - status
      - preview_content
      - subscriber_content
      - scheduled_for
      - published_at
      - created_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
        preview_content:
          type: string
        subscriber_content:
          type: string
        scheduled_for: *id003
        published_at: *id003
        created_at:
          type: string
          format: date-time
    WriterPosts:
      type: object
      required:
      - posts
      - subscriber_count
      properties:
        posts:
          type: array
          items:
            $ref: '#/components/schemas/WriterPost'
        subscriber_count:
          type: integer
          description: Readers with an active allocation to this writer.
    WriterSummary:
      type: object
      required:
      - id
      - display_name
      - first_name
      - last_name
      - avatar_url
      - created_at
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        first_name: *id002
        last_name: *id002
        avatar_url: *id002
        created_at:
          type: string
          format: date-time
    FollowState:
      type: object
      required:
      - following
      properties:
        following:
          type: boolean
    WritersList:
      type: object
      required:
      - writers
      - total
      properties:
        writers:
          type: array
          items:
            $ref: '#/components/schemas/WriterSummary'
        total:
          type: integer
    WriterDetail:
      type: object
      required:
      - id
      - display_name
      - first_name
      - last_name
      - avatar_url
      - created_at
      - subscriber_post_count
      - posts
      - is_following
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        first_name: *id002
        last_name: *id002
        avatar_url: *id002
        created_at:
          type: string
          format: date-time
        subscriber_post_count:
          type: integer
        posts:
          type: array
          items:
            $ref: '#/components/schemas/PostView'
        is_following:
          type: boolean
          description: Whether the (optional) authenticated reader follows this writer.
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from POST /api/v1/auth/login.